fix: move app server models to single file
