--- src/app/shared/models/fcmNotifications.ts
@@ -1,5 +1,0 @@
-import { FcmNotificationDto } from './fcmNotificationDto'
-
-export interface FcmNotifications {
-  notifications?: Array<FcmNotificationDto>
-}
