--- src/app/shared/models/fcmUsers.ts
@@ -1,5 +1,0 @@
-import { FcmUserDto } from './fcmUserDto'
-
-export interface FcmUsers {
-  users?: Array<FcmUserDto>
-}
