--- src/app/shared/models/models.ts
@@ -1,6 +1,0 @@
-export * from './fcmNotificationDto'
-export * from './fcmNotifications'
-export * from './fcmUserDto'
-export * from './fcmUsers'
-export * from './radarProjectDto'
-export * from './radarProjects'
