--- src/app/shared/models/radarProjectDto.ts
@@ -1,4 +1,0 @@
-export interface RadarProjectDto {
-  id?: number
-  projectId: string
-}
