--- src/app/shared/models/radarProjects.ts
@@ -1,5 +1,0 @@
-import { RadarProjectDto } from './radarProjectDto'
-
-export interface RadarProjects {
-  projects?: Array<RadarProjectDto>
-}
